(parser) Refactor and separate canonical chain discovery
